@@ -13,49 +26,63 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "^7.2.2",
-    "@angular/core": "^7.2.2",
-    "@angular/forms": "^7.2.2",
-    "@angular/http": "^7.2.2",
-    "@angular/platform-browser": "^7.2.2",
-    "@angular/platform-browser-dynamic": "^7.2.2",
-    "@angular/router": "^7.2.2",
-    "@auth0/angular-jwt": "^2.1.0",
+    "@angular/common": "^10.0.11",
+    "@angular/compiler": "^10.0.11",
+    "@angular/core": "^10.0.11",
+    "@angular/forms": "^10.0.11",
+    "@angular/localize": "^10.0.11",
+    "@angular/platform-browser": "^10.0.11",
+    "@angular/platform-browser-dynamic": "^10.0.11",
+    "@angular/pwa": "^0.1000.6",
+    "@angular/router": "^10.0.11",
+    "@angular/service-worker": "^10.0.11",
+    "@auth0/angular-jwt": "^5.0.1",
+    "@capacitor/android": "2.4.0",
+    "@capacitor/core": "^2.4.0",
+    "@capacitor/ios": "^2.4.0",
     "@ionic-native/core": "^5.0.0",
     "@ionic-native/splash-screen": "^5.0.0",
     "@ionic-native/status-bar": "^5.0.0",
-    "@ionic/angular": "^4.1.0",
+    "@ionic/angular": "^5.0.1",
     "@ionic/storage": "^2.2.0",
-    "core-js": "^2.5.4",
-    "rxjs": "~6.5.1",
-    "tslib": "^1.9.0",
-    "zone.js": "~0.8.29"
+    "@ngx-translate/core": "^13.0.0",
+    "@ngx-translate/http-loader": "^6.0.0",
+    "chart.js": "^2.8.0",
+    "core-js": "^3.6.5",
+    "firebase": "^7.1.0",
+    "qrcode": "^1.4.2",
+    "rxjs": "^6.6.2",
+    "tslib": "^2.0.0",
+    "zone.js": "~0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "~0.13.8",
-    "@angular-devkit/build-angular": "~0.13.8",
-    "@angular-devkit/core": "~7.3.8",
-    "@angular-devkit/schematics": "~7.3.8",
-    "@angular/cli": "~7.3.8",
-    "@angular/compiler": "~7.2.2",
-    "@angular/compiler-cli": "~7.2.2",
-    "@angular/language-service": "~7.2.2",
-    "@ionic/angular-toolkit": "~1.5.1",
-    "@types/jasmine": "~2.8.8",
+    "@angular-builders/custom-webpack": "^10.0.0",
+    "@angular-devkit/architect": "~0.1000.6",
+    "@angular-devkit/build-angular": "^0.1000.6",
+    "@angular-devkit/core": "~10.0.6",
+    "@angular-devkit/schematics": "~10.0.6",
+    "@angular/cli": "~10.0.6",
+    "@angular/compiler-cli": "^10.0.11",
+    "@angular/language-service": "^10.0.11",
+    "@capacitor/cli": "^1.3.0",
+    "@ionic/angular-toolkit": "^2.3.1",
+    "@types/chart.js": "^2.8.7",
+    "@types/googlemaps": "^3.37.7",
+    "@types/jasmine": "^3.5.12",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~12.0.0",
-    "codelyzer": "~4.5.0",
-    "jasmine-core": "~2.99.1",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.1.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.4.0",
-    "ts-node": "~8.1.0",
-    "tslint": "~5.17.0",
-    "typescript": "~3.1.6"
+    "@types/node": "^14.6.0",
+    "codelyzer": "^6.0.0",
+    "jasmine-core": "~3.5.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "^5.1.1",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-jasmine": "^4.0.1",
+    "karma-jasmine-html-reporter": "^1.5.0",
+    "protractor": "~7.0.0",
+    "ts-node": "~7.0.0",
+    "tslint": "~6.1.0",
+    "typescript": "^3.9.7"
   },
   "description": "An Ionic project"
 }
