@@ -3,6 +3,7 @@ import { Routes, RouterModule } from '@angular/router';
 
 import { LoginPage } from './login.page';
 import { SharedUsuariosModule } from '../shared-usuarios/shared-usuarios.module';
+import { UxHeaderLoginComponent } from './ux-header-login/ux-header-login.component';
 
 const routes: Routes = [
   {
@@ -12,10 +13,7 @@ const routes: Routes = [
 ];
 
 @NgModule({
-  imports: [
-    SharedUsuariosModule,
-    RouterModule.forChild(routes)
-  ],
-  declarations: [LoginPage]
+  imports: [SharedUsuariosModule, RouterModule.forChild(routes)],
+  declarations: [LoginPage, UxHeaderLoginComponent]
 })
 export class LoginPageModule {}
