@@ -5,6 +5,7 @@ import { AuthFormComponent } from './auth-form.component';
 import { ReactiveFormsModule } from '@angular/forms';
 import { IonicModule } from '@ionic/angular';
 import { By } from '@angular/platform-browser';
+import { TranslateModule } from '@ngx-translate/core';
 
 describe('AuthFormComponent', () => {
   let component: AuthFormComponent;
@@ -13,19 +14,21 @@ describe('AuthFormComponent', () => {
     valid: {
       email: 'email@email.com',
       pass: 'asdfF1',
-      rPass: 'asdfF1'
+      rCode: 'asd123',
+      rTOS: true
     },
     invalid: {
       email: 'fdaas',
       pass: 'dsfaaa',
-      rPass: 'dsfaa'
+      rCode: '',
+      rTOS: false
     }
   };
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
       declarations: [AuthFormComponent],
-      imports: [ReactiveFormsModule, IonicModule],
+      imports: [TranslateModule.forRoot(), ReactiveFormsModule, IonicModule],
       schemas: [CUSTOM_ELEMENTS_SCHEMA]
     }).compileComponents();
   }));
@@ -43,17 +46,18 @@ describe('AuthFormComponent', () => {
   it('should call handleSubmit on submit event, valid form', () => {
     component.form.get('email').setValue(formData.valid.email);
     component.form.get('password').setValue(formData.valid.pass);
-    component.form.get('repeat_password').setValue(formData.valid.pass);
     const spy = spyOn(component, 'handleSubmit');
     fixture.detectChanges();
-    fixture.debugElement.query(By.css('form')).triggerEventHandler('ngSubmit', null);
+    fixture.debugElement
+      .query(By.css('form'))
+      .triggerEventHandler('ngSubmit', null);
     expect(spy).toHaveBeenCalledTimes(1);
   });
 
   it('should call emit on send, valid form', () => {
     component.form.get('email').setValue(formData.valid.email);
     component.form.get('password').setValue(formData.valid.pass);
-    component.form.get('repeat_password').setValue(formData.valid.pass);
+    component.form.get('tos').setValue(formData.valid.rTOS);
     fixture.detectChanges();
     const spy = spyOn(component.send, 'emit');
     component.handleSubmit();
@@ -70,35 +74,53 @@ describe('AuthFormComponent', () => {
     it('form should be invalid when fields are empty', async () => {
       component.form.get('email').setValue('');
       component.form.get('password').setValue('');
-      component.form.get('repeat_password').setValue('');
+      component.form.get('tos').setValue(false);
       expect(component.form.valid).toBeFalsy();
     });
 
     it('form should be invalid when email field are not a valid email', async () => {
       component.form.get('email').setValue(formData.invalid.email);
       component.form.get('password').setValue(formData.valid.pass);
-      component.form.get('repeat_password').setValue(formData.valid.pass);
-      expect(component.form.valid).toBeFalsy();
-    });
-
-    it('form should be invalid when passwords not match', async () => {
-      component.form.get('email').setValue(formData.valid.email);
-      component.form.get('password').setValue(formData.valid.pass);
-      component.form.get('repeat_password').setValue(formData.invalid.pass);
+      component.form.get('tos').setValue(formData.valid.rTOS);
       expect(component.form.valid).toBeFalsy();
     });
 
     it('form should be invalid when passwords are not valid', async () => {
       component.form.get('email').setValue(formData.valid.email);
       component.form.get('password').setValue(formData.invalid.pass);
-      component.form.get('repeat_password').setValue(formData.invalid.pass);
       expect(component.form.valid).toBeFalsy();
     });
 
-    it('form should be valid', async () => {
+    it('form should be invalid when TOS are not accepted', async () => {
       component.form.get('email').setValue(formData.valid.email);
       component.form.get('password').setValue(formData.valid.pass);
-      component.form.get('repeat_password').setValue(formData.valid.pass);
+      component.form.get('tos').setValue(formData.invalid.rTOS);
+      expect(component.form.valid).toBeFalsy();
+    });
+
+    it('form should be invalid when manual_referral is true and referral_code is empty', async () => {
+      component.form.get('email').setValue(formData.valid.email);
+      component.form.get('password').setValue(formData.valid.pass);
+      component.form.get('tos').setValue(formData.valid.rTOS);
+      component.form.get('manual_referral').setValue(true);
+      component.form.get('referral_code').setValue(formData.invalid.rCode);
+      expect(component.form.valid).toBeFalsy();
+    });
+
+    it('form should be valid when manual_referral is true and referral_code is valid', async () => {
+      component.form.get('email').setValue(formData.valid.email);
+      component.form.get('password').setValue(formData.valid.pass);
+      component.form.get('tos').setValue(formData.valid.rTOS);
+      component.form.get('manual_referral').setValue(true);
+      component.form.get('referral_code').setValue(formData.valid.rCode);
+      expect(component.form.valid).toBeTruthy();
+    });
+
+    it('form should be valid when manual_referral is False', async () => {
+      component.form.get('email').setValue(formData.valid.email);
+      component.form.get('password').setValue(formData.valid.pass);
+      component.form.get('tos').setValue(formData.valid.rTOS);
+      component.form.get('manual_referral').setValue(false);
       expect(component.form.valid).toBeTruthy();
     });
   });
