@@ -3,9 +3,9 @@ import { CRUD } from 'src/app/shared/services/crud/crud';
 import { CrudService } from 'src/app/shared/services/crud/crud.service';
 import { CustomHttpService } from 'src/app/shared/services/custom-http/custom-http.service';
 import { Observable } from 'rxjs';
-import { API_URL } from 'src/app/config/app-constants.config';
 import { AuthService } from '../auth/auth.service';
 import { tap } from 'rxjs/operators';
+import { environment } from 'src/environments/environment';
 
 @Injectable({
   providedIn: 'root'
@@ -23,22 +23,65 @@ export class ApiUsuariosService {
     this.crud = this.crudService.getEndpoints(this.entity);
   }
 
-  emailValidation(emailValidationToken: string, uidb64: string): Observable<any> {
-    return this.http.post(`${API_URL}/${this.entity}/email-validation/`, {
-      token: emailValidationToken,
-      uidb64
-    });
+  emailValidation(
+    emailValidationToken: string,
+    uidb64: string
+  ): Observable<any> {
+    return this.http.post(
+      `${environment.apiUrl}/${this.entity}/email_validation`,
+      {
+        token: emailValidationToken,
+        uidb64
+      }
+    );
   }
 
   sendEmailValidation(uidb64: string): Observable<any> {
-    return this.http.post(`${API_URL}/${this.entity}/send-email-validation/`, {
-      uidb64
-    });
+    return this.http.post(
+      `${environment.apiUrl}/${this.entity}/send_email_validation`,
+      {
+        uidb64
+      }
+    );
   }
 
   login(data: any): Observable<any> {
     return this.http
-      .post(`${API_URL}/${this.entity}/login/`, data)
+      .post(`${environment.apiUrl}/${this.entity}/login`, data)
       .pipe(tap(response => this.authService.handleLoginResponse(response)));
+  }
+
+  resetPassword(data: any): Observable<any> {
+    return this.http.post(
+      `${environment.apiUrl}/${this.entity}/reset_password`,
+      data
+    );
+  }
+
+  sendResetPasswordEmail(data: any): Observable<any> {
+    return this.http.post(
+      `${environment.apiUrl}/${this.entity}/send_reset_password_email`,
+      data
+    );
+  }
+
+  changePassword(data: any): Observable<any> {
+    return this.http.post(
+      `${environment.apiUrl}/${this.entity}/change_password`,
+      data
+    );
+  }
+
+  getUser(): Observable<any> {
+    return this.http.get(`${environment.apiUrl}/${this.entity}/get_user`);
+  }
+
+  status(loading = true): Observable<any> {
+    return this.http.get(
+      `${environment.apiUrl}/${this.entity}/status`,
+      undefined,
+      undefined,
+      loading
+    );
   }
 }
