@@ -1,15 +1,21 @@
 import { NgModule } from '@angular/core';
 import { AuthFormComponent } from './components/auth-form/auth-form.component';
 import { SharedModule } from 'src/app/shared/shared.module';
+import { ResetPasswordFormComponent } from './components/reset-password-form/reset-password-form.component';
+import { PasswordChangeFormComponent } from './components/password-change-form/password-change-form.component';
 
 @NgModule({
-  declarations: [AuthFormComponent],
-  imports: [
-    SharedModule
+  declarations: [
+    AuthFormComponent,
+    ResetPasswordFormComponent,
+    PasswordChangeFormComponent
   ],
+  imports: [SharedModule],
   exports: [
     SharedModule,
-    AuthFormComponent
+    AuthFormComponent,
+    ResetPasswordFormComponent,
+    PasswordChangeFormComponent
   ]
 })
-export class SharedUsuariosModule { }
+export class SharedUsuariosModule {}
