@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { CustomHttpService } from '../custom-http/custom-http.service';
 import { CRUD } from './crud';
-import { API_URL } from 'src/app/config/app-constants.config';
+import { environment } from 'src/environments/environment';
 
 @Injectable({
   providedIn: 'root'
@@ -11,12 +11,12 @@ export class CrudService {
 
   getEndpoints(entity: string): CRUD {
     return {
-      create: (data: any) => this.http.post(`${API_URL}/${entity}/`, data),
-      get: (id: any) => this.http.get(`${API_URL}/${entity}/${id}/`),
+      create: (data: any) => this.http.post(`${environment.apiUrl}/${entity}/`, data),
+      get: (id?: any) => this.http.get(`${environment.apiUrl}/${entity}/${id || ''}`),
       update: (data: any, id?: any) =>
-        this.http.put(`${API_URL}/${entity}/${data.id || id}/`, data),
-      delete: (id: any) => this.http.delete(`${API_URL}/${entity}/${id}/`),
-      getAll: (options?: any) => this.http.get(`${API_URL}/${entity}/`, '', options)
+        this.http.put(`${environment.apiUrl}/${entity}/${data.id || id || ''}`, data),
+      delete: (id: any) => this.http.delete(`${environment.apiUrl}/${entity}/${id}`),
+      getAll: (options?: any) => this.http.get(`${environment.apiUrl}/${entity}/`, '', options)
     };
   }
 }
