@@ -1,11 +1,14 @@
 import { TestBed } from '@angular/core/testing';
 
 import { LoadingService } from './loading.service';
+import { TranslateModule } from '@ngx-translate/core';
 
 describe('LoadingService', () => {
   let service: LoadingService;
   beforeEach(() => {
-    TestBed.configureTestingModule({});
+    TestBed.configureTestingModule({
+      imports: [TranslateModule.forRoot()]
+    });
     service = TestBed.get(LoadingService);
   });
 
