@@ -1,39 +1,47 @@
 import { Injectable } from '@angular/core';
 import { LoadingController } from '@ionic/angular';
 import { LoadingOptions } from '@ionic/core';
+import { TranslateService } from '@ngx-translate/core';
+import { CONFIG } from 'src/app/config/app-constants.config';
 
 @Injectable({
   providedIn: 'root'
 })
 export class LoadingService {
-
   private message: string;
 
-  private defaultMessage = 'Cargando...';
+  private defaultMessage = CONFIG.loadingService.defaultMessage;
 
   private loading: HTMLIonLoadingElement;
 
   private isVisible: boolean;
   private isEnabled: boolean;
 
-  constructor(private loadingController: LoadingController) {}
+  constructor(
+    private loadingController: LoadingController,
+    private translate: TranslateService
+  ) {}
 
   async show(options: LoadingOptions = {}) {
+    options.cssClass = 'ux-loading';
+    options.message = '';
+    options.spinner = null;
     if (!this.isVisible && this.isEnabled) {
       this.isVisible = true;
-      options = { message: this.message || this.defaultMessage, ...options };
+      const msg = this.message || this.defaultMessage;
+      options = { message: this.translate.instant(msg), ...options };
       this.loading = await this.loadingController.create(options);
       if (this.isVisible) {
-        this.loading.present();
+        await this.loading.present();
       }
     }
   }
 
-  dismiss() {
+  async dismiss() {
     if (this.isVisible) {
       this.isVisible = false;
       if (this.loading) {
-        this.loading.dismiss();
+        await this.loading.dismiss();
       }
       this.message = null;
     }
